Github workflow

+++ tests/setup.js
@@ -1,16 +1,15 @@
 import database from "../src/database/index.js";
-import User from "../src/app/models/User.js";
 
 const WAIT = ms => new Promise(r => setTimeout(r, ms));
 
-async function waitForDatabase() {
+export default async () => {
     let attempts = 10;
 
     while (attempts > 0) {
         try {
             await database.connection.authenticate();
             console.log("📌 Banco disponível para testes");
-            return;
+            break;
         } catch {
             console.log("⏳ Aguardando banco subir...");
             attempts--;
@@ -18,17 +17,9 @@ async function waitForDatabase() {
         }
     }
 
-    throw new Error("❌ Banco de testes não respondeu a tempo");
-}
+    if (attempts === 0) {
+        throw new Error("❌ Banco não respondeu");
+    }
 
-export default async () => {
-    await waitForDatabase();
     await database.connection.sync({ force: true });
-
-    await User.create({
-        name: "Davi",
-        email: "davi@gmail.com",
-        password: "12345678",
-        role: "ADMIN",
-    });
 };
